requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.1.7
django-cors-headers 4.7.0
djangorestframework 3.15.2
pip                 22.3
psycopg             3.2.6
psycopg-binary      3.2.6
PyJWT               2.10.1
python-decouple     3.8
setuptools          65.5.0
sqlparse            0.5.3
typing_extensions   4.13.1
tzdata              2025.2
